feat: 통화를 Money·ExchangedMoney 값 객체로 모델링하고 환율 변환 파이프라인 도입
CURRENCY.md 의 다중 통화 설계를 코드로 구현한다. 요금을 원 통화 원가와 KRW 환산가로 함께 담고,
정규화 단계에서 환율을 적용해 환산한다.

- Money{amount, currency} 값 객체(음수 불가) + ExchangedMoney{original, converted} 로 각 금액이
  자기 통화를 들고 다니게 한다 — converted 는 언제나 KRW, original 은 원 통화임이 타입으로 드러난다.
- Price{exchangeRate, total, averageNightly}: 환율은 총액·평균이 공유하므로 Price 수준에 한 벌만 둔다.
- ExchangeRateProvider 포트 + FixedExchangeRateProvider(통화 무관 1 반환) — 전 공급사 KRW 인 현재는
  no-op 스텁이고, 외화 공급사 추가 시 이 포트를 실제 환율 구현으로 교체한다.
- 어댑터는 원 통화 금액을 BigDecimal grossTotalAmount 로 담기만 하고, 정규화가 Currency·환율로 표준
  Price 를 조립한다. currency 를 String→java.util.Currency 로, validCurrency 를 ISO 파싱 검증으로 강화.
- PriceResponse 는 기존 KRW Long 필드를 환산가로 유지하고 원가·환율 필드를 가산적으로 추가 — KRW
  공급사에서 환율 1·환산가=원가로 자연 degrade 하여 기존 소비자 계약을 깨지 않는다.
- CURRENCY.md 를 구현된 설계(Money/ExchangedMoney, 스텁 환율)에 맞춰 갱신, JOURNAL 에 결정 과정 기록.

+++ src/main/kotlin/com/staysync/supplier/SupplierModels.kt
@@ -5,6 +5,7 @@ import io.netty.channel.ConnectTimeoutException
 import io.netty.handler.timeout.ReadTimeoutException
 import org.springframework.core.codec.DecodingException
 import org.springframework.web.reactive.function.client.WebClientResponseException
+import java.math.BigDecimal
 import java.time.LocalDate
 import java.util.concurrent.TimeoutException
 
@@ -37,7 +38,8 @@ data class SupplierRoomType(
  * 재고·요금 조회 결과 1건 — 공급사 코드 기반의 숙박 상품 (숙소 × 객실 타입 단위).
  *
  * 요금은 두 공급사 교집합인 "세금 포함 총액(gross)"으로 이미 통일해 담는다
- * (A: Σ(nightlyRate + taxAmount), B: totalPrice).
+ * (A: Σ(nightlyRate + taxAmount), B: totalPrice). 원 통화 금액이라 BigDecimal 로 담고, 표준 KRW 로의
+ * 환산은 정규화(검색 계층)가 환율을 적용해 수행한다 (docs/CURRENCY.md).
  */
 data class SupplierStayProduct(
     val supplierPropertyCode: String,
@@ -47,7 +49,7 @@ data class SupplierStayProduct(
     val maxOccupancy: Int,
     val breakfastIncluded: Boolean,
     val currency: String,
-    val grossTotalAmount: Long,
+    val grossTotalAmount: BigDecimal,
     val remainingByDate: Map<LocalDate, Int>,
 )
